+++ src/timeline.js
@@ -498,7 +498,13 @@ export default class Timeline extends React.Component {
      *
      * @type { undefined | boolean}
      */
-    showZommShortcuts: PropTypes.bool
+    showZoomShortcuts: PropTypes.bool,
+
+    /**
+     *
+     * @type {boolean | (() => boolean)}
+     */
+    zoomEnabled: PropTypes.oneOfType([PropTypes.bool, PropTypes.func])
   };
 
   static defaultProps = {
@@ -549,7 +555,8 @@ export default class Timeline extends React.Component {
     zIndexFunction() {
       return 3;
     },
-    showZommShortcuts: false
+    showZoomShortcuts: false,
+    zoomEnabled: true
   };
 
   /**
@@ -760,8 +767,14 @@ export default class Timeline extends React.Component {
     }
   }
 
+  isTouchDevice() {
+    return 'ontouchstart' in window || navigator.maxTouchPoints > 0 || navigator.msMaxTouchPoints > 0;
+  }
+
   wheelHandler(e) {
-    if (e.ctrlKey) {
+    let zoomEnabled =
+      (typeof this.props.zoomEnabled === `function` && this.props.zoomEnabled()) || this.props.zoomEnabled;
+    if (e.ctrlKey && zoomEnabled) {
       let target = e.target;
       while (target) {
         if (target.className.includes(`rct9k-grid rct9k-grid-id-${this.props.componentId}`)) {
@@ -1323,7 +1336,12 @@ export default class Timeline extends React.Component {
       if (this.getDragToCreateMode() && this.props.onDragToCreateEnded) {
         // get avaible itemIndex and call the onDragToCreateEnded
         const itemIndex = Math.max(...Object.keys(this.itemRowMap)) + 1;
-        this.props.onDragToCreateEnded({groupIndex: topRowNumber, itemIndex, itemStart: startTime, itemEnd: endTime});
+        this.props.onDragToCreateEnded({
+          groupIndex: topRowNumber,
+          itemIndex,
+          itemStart: startTime,
+          itemEnd: endTime
+        });
         this.setDragToCreateMode(false);
       }
 
@@ -1509,6 +1527,9 @@ export default class Timeline extends React.Component {
         .resizable({
           allowFrom: selectedItemSelector,
           edges: {left: true, right: true, bottom: false, top: false},
+          // the default from lib was 20 but when we have a small item we cannot move it,
+          // because the size of resize cursor is to big and the move cursor doesn't appear
+          margin: 2,
           ...this.props.interactOptions.draggable
         })
         .on('resizestart', e => {
@@ -1674,8 +1695,7 @@ export default class Timeline extends React.Component {
       let itemKey = target.getAttribute('data-item-index');
       itemKey = isNaN(Number(itemKey)) ? itemKey : Number(itemKey);
       itemCallback && itemCallback(e, itemKey);
-      // window.ontouchstart added to checks is we are on mobile
-      if (e.type == 'mousedown' || (window.ontouchstart && e.type == 'tap')) {
+      if (e.type == 'mousedown' || (this.isTouchDevice() && e.type == 'tap')) {
         // Calculate new selection by delegating to selection component
         this._selectionHolder.addRemoveItems([itemKey], e);
       }
@@ -1691,18 +1711,19 @@ export default class Timeline extends React.Component {
       let snappedClickedTime = timeSnap(clickedTime, this.getTimelineSnap() * 60);
       rowCallback && rowCallback(e, row, clickedTime, snappedClickedTime);
 
-      if (e.type == 'mousedown' || (window.ontouchstart && e.type == 'tap')) {
+      if (e.type == 'mousedown' || (this.isTouchDevice() && e.type == 'tap')) {
         this._selectionHolder.addRemoveItems([], e);
       }
     }
 
     if (e.type == 'contextmenu') {
       // right click => open CM
-      this.setState({openedContextMenuCoordinates: {x: e.clientX, y: e.clientY}});
+      const {clientX, clientY} = e;
+      this.setState({openedContextMenuCoordinates: {x: clientX, y: clientY}});
       this.setState({openedContextMenuRow: Number(row)});
       this.setState({
         openedContextMenuTime: getTimeAtPixel(
-          e.clientX - this.getGanttLeftOffset(),
+          clientX - this.getGanttLeftOffset(),
           this.getStartDate(),
           this.getEndDate(),
           this.getTimelineWidth(),
@@ -1714,7 +1735,7 @@ export default class Timeline extends React.Component {
         // If a drag in progress, right click only cancels the current drag but keeps the drag to create mode
         this.setDragToCreateMode(false);
       }
-    } else if (e.type == 'click' || (window.ontouchstart && e.type == 'tap')) {
+    } else if (e.type == 'click' || (this.isTouchDevice() && e.type == 'tap')) {
       this.setState({openedContextMenuCoordinates: undefined});
       this.setDragToCreateMode(false);
     }
@@ -1761,12 +1782,18 @@ export default class Timeline extends React.Component {
             this.selecting = false;
             return this._handleItemRowEvent(e, onItemHover, null);
           }}
-          onMouseLeave={e => {
+          onMouseOut={e => {
             this.selecting = false;
             return this._handleItemRowEvent(e, onItemLeave, null);
           }}
-          onContextMenu={e => this._handleItemRowEvent(e, this.props.onItemContextClick, this.props.onRowContextClick)}
-          onDoubleClick={e => this._handleItemRowEvent(e, this.props.onItemDoubleClick, this.props.onRowDoubleClick)}>
+          onContextMenu={e => {
+            this.selecting = false;
+            this._handleItemRowEvent(e, this.props.onItemContextClick, this.props.onRowContextClick);
+          }}
+          onDoubleClick={e => {
+            this.selecting = false;
+            this._handleItemRowEvent(e, this.props.onItemDoubleClick, this.props.onRowDoubleClick);
+          }}>
           {rowItemsRenderer(
             itemsInRow,
             this.getStartDate(),
@@ -1975,12 +2002,12 @@ export default class Timeline extends React.Component {
     // directly on these components was not possible easily (It needed to extend this Grid adding by adding a wrapper div on which to add the touch handlers)
     // so we added the handler on the parent component and check inside of it to exclude other children like the scrollbar, or the timebar
     if (
+      !this.state.hasHorizontalScrollbar ||
       e.target.classList.contains('rct9k-horizontal-scrollbar-outter') ||
       e.target.classList.contains('rct9k-timebar-item')
     ) {
       return;
     }
-
     const touch = e.touches[0];
     this.setState({touchPositionX: touch.clientX});
   }
@@ -1991,7 +2018,7 @@ export default class Timeline extends React.Component {
    * @param {*} e
    */
   onTouchMove(e) {
-    if (this.state.touchPositionX != undefined) {
+    if (this.state.touchPositionX != undefined && this._scrollbar != undefined) {
       const touch = e.touches[0];
       const deltaX = touch.clientX - this.state.touchPositionX;
       this.setState({touchPositionX: touch.clientX});
@@ -2005,8 +2032,8 @@ export default class Timeline extends React.Component {
    * Toghether with `onTouchStart` and `onTouchMove` implements the horizontal scroll by dragging the gantt diagram on mobile devices
    *
    */
-  onTouchEnd() {
-    this.setState({touchPositionX: undefined});
+  onTouchEnd(e) {
+    this.state.touchPositionX && this.setState({touchPositionX: undefined});
   }
 
   /**
@@ -2135,7 +2162,7 @@ export default class Timeline extends React.Component {
         }
       });
     }
-    if (this.props.showZommShortcuts) {
+    if (this.props.showZoomShortcuts) {
       let that = this;
       actions.push({
         label: ZOOM_IN_ACTION_LABEL,